Finalise nested routing
