+++ admin_ui/app/App.js
@@ -6,11 +6,13 @@ import Store from './Store';
 /*
   We want both the Login view and the Dashboard view to be part of the main bundle
   (and thus appear immediatelly to the user once the bundle is loaded)
-  DashboardPage contains a nested <Router> which loads all other pages using
-  react-loadable.
+  DashboardSection contains a nested <Router> which loads all
+  other pages using react-loadable
  */
 import LoginPage from './session/LoginPage';
-import DashboardPage from './dashboard/DashboardPage';
+import DashboardSection from './dashboard/DashboardSection';
+
+
 
 /**
  * The top level component / wrapper
@@ -19,8 +21,8 @@ const App = () => {
   return (
     <Store.Provider>
       <Router>
+        <DashboardSection path="/*" />
         <LoginPage path="/login" />
-        <DashboardPage path="/" />
       </Router>
     </Store.Provider>
   );
